Changes to TrackingException API to return indication of site-wide consent
The storeSiteSpeciificTrackingException now returns a Promise resolving to a "TrackingExceptionResult" object that contains a boolean "siteWide", set to true if a site-wide exception has been registered.

For consistency and simplicity of implementation the confirm calls (both confirmSiteSpecificTrackingException and confirmWebWideTrackingException now also return a TrackingExcpetionResult object with a boolean "exists" property indicating existence, rather than resolving to a boolean which would have been inconsistent. storeWebWideTrackingException, removeSiteSpecificTrackingException and removeWebWideTrackingException still resolve to a void.