Fix the wording introduced for issue #21 so that it doesn't conflict with previous consensus regarding TSR validity and the defaults for controller and policy, all of which were already defined (where it belongs).
Move definition of user-granted exception to its defining section and
rewrite the confusing overview,

Fix an old reference to window.doNotTrack (now on Navigator).
Remove informative distinctions when that is invalid or obvious.
Remove notes that are redundant to the normative content.
Use defined term "target resource" instead of request target or
other variations on the theme.

Editorial fixes due to changes in how respec handles crossrefs.