Combines web-wide and site-specific APIs
Now there are only 3 API functions, storeTrackingException, removeTrackingException and confirmTrackingException, with a new property in the propertybags determining if the call is web-wide or site-specific.
All calls now reject rather than throwing an exception. It uses DOMException.SYNTAX_ERR for wrong parameters and DOMException.SECURITY_ERR for cookie domain errors.